Refactor 인벤토리 조작 이동방식 변경
x,y 인덱스를 이용한 그리드 형식에서 좌표 및 거리계산 형식으로 변경.

+++ Assets/02_Scripts/Utils/Utils.cs
@@ -854,7 +854,7 @@ namespace Default
 
         public static int GetRatioIntByInt(int value, int from, int to)
         {
-            int realInt = Mathf.RoundToInt((float)value / (float)(from - 1) * (to - 1));
+            int realInt = Mathf.RoundToInt((float)value / (from - 1) * (to - 1));
             return Mathf.Clamp(realInt, 0, to - 1);
         }
     }
